+++ packages/embed/e2e/spec/functional/popover.cy.ts
@@ -30,7 +30,7 @@ function testPopover(path: string, title: string) {
     it('should pass options as query param', () => {
       cy.get('.tf-v1-popover-wrapper iframe')
         .invoke('attr', 'src')
-        .should('contain', 'typeform-embed=popup-popover&typeform-source=localhost&typeform-medium=demo-test')
+        .should('contain', 'buildai-embed=popup-popover&buildai-source=localhost&buildai-medium=demo-test')
     })
 
     it('should pass hidden fields as hash', () => {