+++ packages/embed/src/utils/build-iframe-src.spec.ts
@@ -14,11 +14,11 @@ describe('build-iframe-src', () => {
       it('should include default url options', () => {
         expect(buildIframeSrc({ formId: 'some-id', type: 'widget', embedId: 'embed-id', options: {} })).toBe(
           'https://form.typeform.com/to/some-id' +
-            '?typeform-embed-id=embed-id' +
-            '&typeform-embed=embed-widget' +
-            '&typeform-source=localhost' +
-            '&typeform-medium=embed-sdk' +
-            '&typeform-medium-version=next'
+            '?buildai-embed-id=embed-id' +
+            '&buildai-embed=embed-widget' +
+            '&buildai-source=localhost' +
+            '&buildai-medium=embed-sdk' +
+            '&buildai-medium-version=next'
         )
       })
     })
@@ -27,11 +27,11 @@ describe('build-iframe-src', () => {
       it('should include the full URL', () => {
         const formId = 'https://custom.example.com/form-id'
         const iframeSrcParams =
-          '?typeform-embed-id=embed-id' +
-          '&typeform-embed=embed-widget' +
-          '&typeform-source=localhost' +
-          '&typeform-medium=embed-sdk' +
-          '&typeform-medium-version=next'
+          '?buildai-embed-id=embed-id' +
+          '&buildai-embed=embed-widget' +
+          '&buildai-source=localhost' +
+          '&buildai-medium=embed-sdk' +
+          '&buildai-medium-version=next'
         expect(buildIframeSrc({ formId, type: 'widget', embedId: 'embed-id', options: {} })).toBe(
           `${formId}${iframeSrcParams}`
         )
@@ -41,11 +41,11 @@ describe('build-iframe-src', () => {
     describe('when passing custom domain', () => {
       it('should prefer it over the default one', () => {
         const iframeSrcParams =
-          '?typeform-embed-id=embed-id' +
-          '&typeform-embed=embed-widget' +
-          '&typeform-source=localhost' +
-          '&typeform-medium=embed-sdk' +
-          '&typeform-medium-version=next'
+          '?buildai-embed-id=embed-id' +
+          '&buildai-embed=embed-widget' +
+          '&buildai-source=localhost' +
+          '&buildai-medium=embed-sdk' +
+          '&buildai-medium-version=next'
         expect(
           buildIframeSrc({
             formId: 'formId',
@@ -60,11 +60,11 @@ describe('build-iframe-src', () => {
       it('should ignore the domain if formID is URL', () => {
         const formId = 'https://custom.example.com/form-id'
         const iframeSrcParams =
-          '?typeform-embed-id=embed-id' +
-          '&typeform-embed=embed-widget' +
-          '&typeform-source=localhost' +
-          '&typeform-medium=embed-sdk' +
-          '&typeform-medium-version=next'
+          '?buildai-embed-id=embed-id' +
+          '&buildai-embed=embed-widget' +
+          '&buildai-source=localhost' +
+          '&buildai-medium=embed-sdk' +
+          '&buildai-medium-version=next'
         expect(
           buildIframeSrc({ formId, type: 'widget', domain: 'foobar.example.net', embedId: 'embed-id', options: {} })
         ).toBe(`${formId}${iframeSrcParams}`)
@@ -78,8 +78,8 @@ describe('build-iframe-src', () => {
         embedId: 'id',
         options: { medium: 'unit-test-medium', source: 'unit-test-source' },
       })
-      expect(src).toMatch('typeform-medium=unit-test-medium')
-      expect(src).toMatch('typeform-source=unit-test-source')
+      expect(src).toMatch('buildai-medium=unit-test-medium')
+      expect(src).toMatch('buildai-source=unit-test-source')
     })
 
     it('should omit false url options', () => {
@@ -117,18 +117,18 @@ describe('build-iframe-src', () => {
       }
       expect(buildIframeSrc({ formId: 'some-id', type: 'widget', embedId: 'embed-id', options })).toBe(
         'https://form.typeform.com/to/some-id' +
-          '?typeform-embed-id=embed-id' +
-          '&typeform-embed=embed-widget' +
-          '&typeform-source=unit-test-source' +
-          '&typeform-medium=unit-test-medium' +
-          '&typeform-medium-version=unit-test-version' +
+          '?buildai-embed-id=embed-id' +
+          '&buildai-embed=embed-widget' +
+          '&buildai-source=unit-test-source' +
+          '&buildai-medium=unit-test-medium' +
+          '&buildai-medium-version=unit-test-version' +
           '&embed-hide-footer=true' +
           '&embed-hide-headers=true' +
           '&embed-opacity=50' +
           '&disable-tracking=true' +
           '&__dangerous-disable-submissions=true' +
-          '&typeform-embed-auto-resize=true' +
-          '&typeform-embed-handle-ending-button-click=true' +
+          '&buildai-embed-auto-resize=true' +
+          '&buildai-embed-handle-ending-button-click=true' +
           '&utm_foo=utm+foo+value&foobar=foobar%26value' +
           '#foo=foo+value&bar=%40bar%26value%3F' +
           '&hubspot_page_name=page+title&hubspot_page_url=http%3A%2F%2Flocalhost%2F'
@@ -145,11 +145,11 @@ describe('build-iframe-src', () => {
 
       expect(buildIframeSrc({ formId: 'some-id', type: 'widget', embedId: 'embed-id', options })).toBe(
         'https://form.typeform.com/to/some-id' +
-          '?typeform-embed-id=embed-id' +
-          '&typeform-embed=embed-widget' +
-          '&typeform-source=unit-test-source' +
-          '&typeform-medium=unit-test-medium' +
-          '&typeform-medium-version=unit-test-version' +
+          '?buildai-embed-id=embed-id' +
+          '&buildai-embed=embed-widget' +
+          '&buildai-source=unit-test-source' +
+          '&buildai-medium=unit-test-medium' +
+          '&buildai-medium-version=unit-test-version' +
           '&disable-tracking=true' +
           '&__dangerous-disable-submissions=true'
       )